Hopefully fixed content loading bug

+++ Sprint0/Sprint0/DelayGame.cs
@@ -54,7 +54,7 @@ namespace Cuphead
             else
             {
                 delayFlag = true;
-                delayFrames = frames; 
+                delayFrames = (int)frames; 
                 frameCount = 0;
             }
             return false;
